+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## unreleased
+* Add `debitNetwork` field to `Transaction` and `TransactionSearchRequest`
+* Add `DebitNetwork` enum to `CreditCard`
+
 ## 3.22.0
 * Add subscriptions to `SepaDirectDebitAccount`
 * Add device data fields to `ThreeDSecureLookupRequest`